debundle: split graph.rs into submodules
graph.rs was a ~2.6k-line module. Split it into cohesive submodules
behind the same public surface; mod.rs is now a thin module index.

  edge          edge kinds and edge-set helpers
  build         graph construction from facts
  owner_graph   owner-graph derivation
  linker_order  linker / source-order computation
  quotient      SCC quotient graph
  tests         module tests

No behavior change; all 116 debundle tests pass, clippy clean.
