+++ devinfra/js/debundle/BUILD.bazel
@@ -360,7 +360,13 @@ rust_library(
         "facts/local_effects.rs",
         "facts/mod.rs",
         "facts/wire.rs",
-        "graph.rs",
+        "graph/build.rs",
+        "graph/edge.rs",
+        "graph/linker_order.rs",
+        "graph/mod.rs",
+        "graph/owner_graph.rs",
+        "graph/quotient.rs",
+        "graph/tests.rs",
         "ids.rs",
         "lib.rs",
         "partition.rs",
